+++ tmp/__init__.py
@@ -1,11 +1,11 @@
-from functools import wraps
 import os
 
-from flask import Flask, abort, g, redirect, render_template, request, url_for
+from flask import Flask, g
 from sqlalchemy import create_engine
 from sqlalchemy.orm import scoped_session, sessionmaker
 
-from .models import Base, User
+from .models import Base
+from .matrix import matrix
 
 
 DATABASE_URI = os.getenv('DATABASE_URI')
@@ -18,24 +18,15 @@ app = Flask(
 )
 
 
+app.register_blueprint(matrix, url_prefix='/matrix')
+
+
 @app.before_first_request
 def initialize():
-    global Session
     engine = create_engine(DATABASE_URI)
-    Session = scoped_session(sessionmaker(engine))
+    app.Session = scoped_session(sessionmaker(engine))
     Base.metadata.create_all(engine)
-    Base.query = Session.query_property()
-
-
-def use_db(f):
-
-    @wraps(f)
-    def wrapped(*args, **kwarg):
-        assert not hasattr(g, 'db')
-        g.db = Session()
-        return f(*args, **kwarg)
-
-    return wrapped
+    Base.query = app.Session.query_property()
 
 
 @app.after_request
@@ -44,47 +35,3 @@ def after_request(response):
         g.db.close()
 
     return response
-
-
-def get_ip():
-    forwarded = request.headers.getlist('X-Forwarded-For')
-    if forwarded:
-        return forwarded[0]
-
-    return request.remote_addr
-
-
-@app.route('/')
-@use_db
-def index():
-    ip = get_ip()
-    user = g.db.query(User).get(ip)
-    if user:
-        # TODO: Make these entry points.
-        if user.choice == 'blue':
-            return redirect('out')
-        elif user.choice == 'red':
-            return redirect('second')
-
-    return render_template('index.html')
-
-
-@app.route('/choice')
-@use_db
-def choice():
-    color = request.args.get('choice')
-    if color not in ('blue', 'red'):
-        return abort(400)
-
-    ip = get_ip()
-
-    user = g.db.query(User).get(ip)
-    if user:
-        return redirect(url_for('index'))
-
-    user = User()
-    user.ip = ip
-    user.choice = color
-    g.db.add(user)
-    g.db.commit()
-    return redirect(url_for('index'))
