+++ tmp/decorators.py
@@ -1,0 +1,14 @@
+from functools import wraps
+
+from flask import current_app, g
+
+
+def use_db(f):
+
+    @wraps(f)
+    def wrapped(*args, **kwarg):
+        assert not hasattr(g, 'db')
+        g.db = current_app.Session()
+        return f(*args, **kwarg)
+
+    return wrapped
