+++ tmp/templates/matrix/index.html
@@ -17,8 +17,8 @@
       </br>Driving you mad. It is this feeling that has brought you to me.
     </br>The [Censored] is everywhere. It is all around us. Even now, in this very room.</p>
     <p class="center">
-      <a role="button" href="{{ url_for('choice', choice='red') }}"><img src="" alt="Red pill" /></a>
-      <a role="button" href="{{ url_for('choice', choice='blue') }}"><img src="" alt="Blue pill" /></a>
+      <a role="button" href="{{ url_for('.choice', choice='red') }}"><img src="" alt="Red pill" /></a>
+      <a role="button" href="{{ url_for('.choice', choice='blue') }}"><img src="" alt="Blue pill" /></a>
     </p>
     <script src="{{ url_for('static', filename='index.js') }}"></script>
   </body>
